+++ apps/wallet/src/ui/app/components/menu/content/AccountsSettings.tsx
@@ -20,6 +20,9 @@ export function AccountsSettings() {
     const backUrl = useNextMenuUrl(true, '/');
     const importPrivateKeyUrl = useNextMenuUrl(true, '/import-private-key');
     const accounts = useAccounts();
+    const isMultiAccountsEnabled = useFeature(
+        FEATURES.WALLET_MULTI_ACCOUNTS
+    ).on;
     const createAccountMutation = useDeriveNextAccountMutation();
     const { on: isLedgerIntegrationEnabled } = useFeature(
         FEATURES.WALLET_LEDGER_INTEGRATION
@@ -41,19 +44,23 @@ export function AccountsSettings() {
                         accountType={type}
                     />
                 ))}
-                <Button
-                    variant="outline"
-                    size="tall"
-                    text="Create New Account"
-                    loading={createAccountMutation.isLoading}
-                    onClick={() => createAccountMutation.mutate()}
-                />
-                <Button
-                    variant="outline"
-                    size="tall"
-                    text="Import Private Key"
-                    to={importPrivateKeyUrl}
-                />
+                {isMultiAccountsEnabled ? (
+                    <>
+                        <Button
+                            variant="outline"
+                            size="tall"
+                            text="Create New Account"
+                            loading={createAccountMutation.isLoading}
+                            onClick={() => createAccountMutation.mutate()}
+                        />
+                        <Button
+                            variant="outline"
+                            size="tall"
+                            text="Import Private Key"
+                            to={importPrivateKeyUrl}
+                        />
+                    </>
+                ) : null}
                 {isLedgerIntegrationEnabled ? (
                     <Button
                         variant="outline"
