+++ apps/wallet/src/ui/app/components/menu/content/index.tsx
@@ -2,9 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { useFeature } from '@growthbook/growthbook-react';
-import { LockedDeviceError } from '@ledgerhq/errors';
 import { useCallback } from 'react';
-import toast from 'react-hot-toast';
 import {
     Navigate,
     Route,
@@ -13,12 +11,8 @@ import {
     useNavigate,
 } from 'react-router-dom';
 
-import { ConnectLedgerModal } from '../../ledger/ConnectLedgerModal';
+import { ConnectLedgerModalContainer } from '../../ledger/ConnectLedgerModalContainer';
 import { ImportLedgerAccounts } from '../../ledger/ImportLedgerAccounts';
-import {
-    LedgerConnectionFailedError,
-    LedgerNoTransportMechanismError,
-} from '../../ledger/LedgerExceptions';
 import { AccountsSettings } from './AccountsSettings';
 import { AutoLockSettings } from './AutoLockSettings';
 import { ExportAccount } from './ExportAccount';
@@ -40,7 +34,6 @@ import type { MouseEvent } from 'react';
 const CLOSE_KEY_CODES: string[] = ['Escape'];
 
 function MenuContent() {
-    const accountsUrl = useNextMenuUrl(true, '/accounts');
     const mainLocation = useLocation();
     const isOpen = useMenuIsOpen();
     const menuUrl = useMenuUrl();
@@ -58,13 +51,11 @@ function MenuContent() {
     );
     useOnKeyboardEvent('keydown', CLOSE_KEY_CODES, handleOnCloseMenu, isOpen);
 
+    const isMultiAccountsEnabled = useFeature(
+        FEATURES.WALLET_MULTI_ACCOUNTS
+    ).on;
     const { on: isLedgerIntegrationEnabled } = useFeature(
         FEATURES.WALLET_LEDGER_INTEGRATION
-    );
-
-    const importLedgerAccountsUrl = useNextMenuUrl(
-        true,
-        '/import-ledger-accounts'
     );
 
     if (!isOpen) {
@@ -77,36 +68,31 @@ function MenuContent() {
                 <MainLocationContext.Provider value={mainLocation}>
                     <Routes location={menuUrl || ''}>
                         <Route path="/" element={<MenuList />} />
-
-                        <Route path="/accounts" element={<AccountsSettings />}>
-                            <Route
-                                path="connect-ledger-modal"
-                                element={
-                                    <ConnectLedgerModal
-                                        onClose={() => {
-                                            navigate(accountsUrl);
-                                        }}
-                                        onError={(error) => {
-                                            navigate(accountsUrl);
-                                            toast.error(
-                                                getLedgerErrorMessage(error)
-                                            );
-                                        }}
-                                        onConfirm={() => {
-                                            navigate(importLedgerAccountsUrl);
-                                        }}
-                                    />
-                                }
-                            />
-                        </Route>
-                        <Route
-                            path="/export/:account"
-                            element={<ExportAccount />}
-                        />
-                        <Route
-                            path="/import-private-key"
-                            element={<ImportPrivateKey />}
-                        />
+                        {isMultiAccountsEnabled ? (
+                            <>
+                                <Route
+                                    path="/accounts"
+                                    element={<AccountsSettings />}
+                                >
+                                    {isLedgerIntegrationEnabled && (
+                                        <Route
+                                            path="connect-ledger-modal"
+                                            element={
+                                                <ConnectLedgerModalContainer />
+                                            }
+                                        />
+                                    )}
+                                </Route>
+                                <Route
+                                    path="/export/:account"
+                                    element={<ExportAccount />}
+                                />
+                                <Route
+                                    path="/import-private-key"
+                                    element={<ImportPrivateKey />}
+                                />
+                            </>
+                        ) : null}
 
                         <Route path="/network" element={<NetworkSettings />} />
                         <Route
@@ -133,14 +119,3 @@ function MenuContent() {
 }
 
 export default MenuContent;
-
-function getLedgerErrorMessage(error: unknown) {
-    if (error instanceof LockedDeviceError) {
-        return 'Your device is locked. Un-lock it and try again.';
-    } else if (error instanceof LedgerConnectionFailedError) {
-        return 'Ledger connection failed.';
-    } else if (error instanceof LedgerNoTransportMechanismError) {
-        return "Your machine doesn't support USB or HID.";
-    }
-    return 'Something went wrong. Try again.';
-}
